Leave existing loggers enabled on fileConfig read.
From the logging documentation at:
https://docs.python.org/3/library/logging.config.html
If logging.config.fileConfig is called, with disable_existing_loggers
not set, then the default behaviour is to remove all non-root loggers
from the hierarchy. This has the side effect of removing all of the
uvicorn logging, so instead we simply set disable_existing_loggers to
False.